Check for fused kernel before inplace update
Summary:
Given an op, with a pair (output buffer, input buffer) from that op, we consider
marking the output buffer as inline. However, if the parent of input buffer and
the current op are going to be fused, then we don't want to mark the output
buffer as inline. This change checks that criteria, and skips inlining if it is
so.

Test Plan:
New unit test "layer_norm_should_not_inplace" runs LayerNorm and checks for no
"in_out" pointers.

Tasks:
Resolves https://github.com/pytorch/pytorch/issues/120217